train/requirements.txt
torch
torchvision
pytorch-lightning
numpy
Pillow
argparse
pycocotools
torchmetrics